+++ R/geom-contour.r
@@ -1,14 +1,15 @@
-#' 2d contours of a 3d surface
+#' 2D contours of a 3D surface
 #'
-#' ggplot2 can not draw true 3d surfaces, but you can use `geom_contour`
-#' and [geom_tile()] to visualise 3d surfaces in 2d. To be a valid
-#' surface, the data must contain only a single row for each unique combination
-#' of the variables mapped to the `x` and `y` aesthetics. Contouring
+#' ggplot2 can not draw true 3D surfaces, but you can use `geom_contour()`,
+#' `geom_contour_filled()`, and [geom_tile()] to visualise 3D surfaces in 2D.
+#' To specify a valid surface, the data must contain `x`, `y`, and `z` coordinates,
+#' and each unique combination of `x` and `y` can appear exactly once. Contouring
 #' tends to work best when `x` and `y` form a (roughly) evenly
 #' spaced grid. If your data is not evenly spaced, you may want to interpolate
-#' to a grid before visualising.
+#' to a grid before visualising, see [geom_density_2d()].
 #'
 #' @eval rd_aesthetics("geom", "contour")
+#' @eval rd_aesthetics("geom", "contour_filled")
 #' @inheritParams layer
 #' @inheritParams geom_point
 #' @inheritParams geom_path
@@ -20,7 +21,7 @@
 #' @seealso [geom_density_2d()]: 2d density contours
 #' @export
 #' @examples
-#' #' # Basic plot
+#' # Basic plot
 #' v <- ggplot(faithfuld, aes(waiting, eruptions, z = density))
 #' v + geom_contour()
 #'
@@ -33,7 +34,7 @@
 #' v + geom_contour_filled()
 #'
 #' # Setting bins creates evenly spaced contours in the range of the data
-#' v + geom_contour(bins = 2)
+#' v + geom_contour(bins = 5)
 #' v + geom_contour(bins = 10)
 #'
 #' # Setting binwidth does the same thing, parameterised by the distance
@@ -95,7 +96,7 @@ geom_contour_filled <- function(mapping = NULL, data = NULL,
     data = data,
     mapping = mapping,
     stat = stat,
-    geom = GeomPolygon,
+    geom = GeomContourFilled,
     position = position,
     show.legend = show.legend,
     inherit.aes = inherit.aes,
@@ -123,3 +124,11 @@ GeomContour <- ggproto("GeomContour", GeomPath,
     alpha = NA
   )
 )
+
+#' @rdname ggplot2-ggproto
+#' @format NULL
+#' @usage NULL
+#' @export
+#' @include geom-polygon.r
+GeomContourFilled <- ggproto("GeomContourFilled", GeomPolygon)
+